advice page data ready
Advice page data ready, I wrote mongo script to get all data needed. I tried to fix  the problem of conflict between angular and template, but it turn out to be too time consuming. So for now, I was using the same approach of getting data and load the page.